backend/requirements.txt
Flask
Flask-Cors
yt-dlp
moviepy
openai-whisper
huggingface-hub
graphviz
